Move clone to Activity again
